feat: factorize task flatener
