+++ src/ui/hooks/useStdoutDimensions.ts
